+++ tests/functional/test_deployments.py
@@ -10,8 +10,7 @@ EXPECTED_HEADERS = deployments_commands.default_headers
 
 
 class TestDeploymentsCreate(object):
-    URL = "https://api.paperspace.io/deployments/createDeployment/"
-    URL_V2 = "https://api.paperspace.io/deployments/v2/createDeployment/"
+    URL = "https://api.paperspace.io/deployments/v2/createDeployment/"
     EXPECTED_HEADERS_WITH_CHANGED_API_KEY = http_client.default_headers.copy()
     EXPECTED_HEADERS_WITH_CHANGED_API_KEY["X-API-Key"] = "some_key"
     BASIC_OPTIONS_COMMAND = [
@@ -22,16 +21,7 @@ class TestDeploymentsCreate(object):
         "--machineType", "G1",
         "--imageUrl", "https://www.latlmes.com/breaking/paperspace-now-has-a-100-bilion-valuation",
         "--instanceCount", "666",
-    ]
-    BASIC_OPTIONS_COMMAND_WITH_USE_VPC_FLAG = [
-        "deployments", "create",
-        "--deploymentType", "tfserving",
-        "--modelId", "some_model_id",
-        "--name", "some_name",
-        "--machineType", "G1",
-        "--imageUrl", "https://www.latlmes.com/breaking/paperspace-now-has-a-100-bilion-valuation",
-        "--instanceCount", "666",
-        "--vpc",
+        "--clusterId", "cluster"
     ]
     BASIC_OPTIONS_COMMAND_WITH_API_KEY = [
         "deployments", "create",
@@ -42,6 +32,7 @@ class TestDeploymentsCreate(object):
         "--imageUrl", "https://www.latlmes.com/breaking/paperspace-now-has-a-100-bilion-valuation",
         "--instanceCount", "666",
         "--apiKey", "some_key",
+        "--clusterId", "cluster",
     ]
     COMMAND_WITH_ALL_OPTIONS = [
         "deployments", "create",
@@ -66,7 +57,6 @@ class TestDeploymentsCreate(object):
         "--authPassword", "some_password",
         "--clusterId", "some_cluster_id",
         "--apiKey", "some_key",
-        "--vpc",
     ]
     COMMAND_WITH_OPTIONS_FILE = ["deployments", "create", "--optionsFile", ]  # path added in test
 
@@ -77,6 +67,7 @@ class TestDeploymentsCreate(object):
         "deploymentType": "TFServing",
         "instanceCount": 666,
         "modelId": u"some_model_id",
+        "cluster": "cluster"
     }
     ALL_OPTIONS_REQUEST = {
         "machineType": u"G1",
@@ -126,23 +117,6 @@ class TestDeploymentsCreate(object):
         assert result.exit_code == 0
 
     @mock.patch("gradient.cli.deployments.deployments_commands.http_client.requests.post")
-    def test_should_send_request_to_api_v2_when_vpc_flag_was_used(self, post_patched):
-        post_patched.return_value = MockResponse(self.RESPONSE_JSON_200, 200, "fake content")
-
-        runner = CliRunner()
-        result = runner.invoke(cli.cli, self.BASIC_OPTIONS_COMMAND_WITH_USE_VPC_FLAG)
-
-        assert result.output == self.EXPECTED_STDOUT, result.exc_info
-        post_patched.assert_called_once_with(self.URL_V2,
-                                             headers=EXPECTED_HEADERS,
-                                             json=self.BASIC_OPTIONS_REQUEST,
-                                             params=None,
-                                             files=None,
-                                             data=None)
-
-        assert result.exit_code == 0
-
-    @mock.patch("gradient.cli.deployments.deployments_commands.http_client.requests.post")
     def test_should_send_proper_data_and_print_message_when_create_deployment_with_all_options(self, post_patched):
         post_patched.return_value = MockResponse(self.RESPONSE_JSON_200)
 
@@ -150,7 +124,7 @@ class TestDeploymentsCreate(object):
         result = runner.invoke(cli.cli, self.COMMAND_WITH_ALL_OPTIONS)
 
         assert result.output == self.EXPECTED_STDOUT, result.exc_info
-        post_patched.assert_called_once_with(self.URL_V2,
+        post_patched.assert_called_once_with(self.URL,
                                              headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
                                              json=self.ALL_OPTIONS_REQUEST,
                                              params=None,
@@ -183,7 +157,7 @@ class TestDeploymentsCreate(object):
         result = runner.invoke(cli.cli, command)
 
         assert result.output == self.EXPECTED_STDOUT, result.exc_info
-        post_patched.assert_called_once_with(self.URL_V2,
+        post_patched.assert_called_once_with(self.URL,
                                              headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
                                              json=self.ALL_OPTIONS_REQUEST,
                                              params=None,
@@ -349,11 +323,9 @@ class TestDeploymentsList(object):
 
 
 class TestStartDeployment(object):
-    URL = "https://api.paperspace.io/deployments/updateDeployment/"
-    URL_V2 = "https://api.paperspace.io/deployments/v2/updateDeployment/"
+    URL = "https://api.paperspace.io/deployments/v2/updateDeployment/"
     COMMAND = ["deployments", "start",
                "--id", "some_id"]
-    COMMAND_WITH_VPC_FLAG = ["deployments", "start", "--id", "some_id", "--vpc"]
     COMMAND_WITH_OPTIONS_FILE = ["deployments", "start", "--optionsFile", ]  # path added in test
     REQUEST_JSON = {"isRunning": True, "id": u"some_id"}
     EXPECTED_STDOUT = "Deployment started\n"
@@ -378,22 +350,6 @@ class TestStartDeployment(object):
         assert result.exit_code == 0
 
     @mock.patch("gradient.cli.deployments.deployments_commands.http_client.requests.post")
-    def test_should_send_request_to_api_v2_when_vpc_flag_was_used(self, post_patched):
-        post_patched.return_value = MockResponse(status_code=204)
-
-        runner = CliRunner()
-        result = runner.invoke(cli.cli, self.COMMAND_WITH_VPC_FLAG)
-
-        assert result.output == self.EXPECTED_STDOUT, result.exc_info
-        post_patched.assert_called_once_with(self.URL_V2,
-                                             headers=EXPECTED_HEADERS,
-                                             json=self.REQUEST_JSON,
-                                             params=None,
-                                             files=None,
-                                             data=None)
-        assert result.exit_code == 0
-
-    @mock.patch("gradient.cli.deployments.deployments_commands.http_client.requests.post")
     def test_should_read_options_from_file(self, post_patched, deployments_start_config_path):
         post_patched.return_value = MockResponse(status_code=204)
         command = self.COMMAND_WITH_OPTIONS_FILE[:] + [deployments_start_config_path]
@@ -402,7 +358,7 @@ class TestStartDeployment(object):
         result = runner.invoke(cli.cli, command)
 
         assert result.output == self.EXPECTED_STDOUT, result.exc_info
-        post_patched.assert_called_once_with(self.URL_V2,
+        post_patched.assert_called_once_with(self.URL,
                                              headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
                                              json=self.REQUEST_JSON,
                                              params=None,
@@ -412,13 +368,9 @@ class TestStartDeployment(object):
 
 
 class TestStopDeployment(object):
-    URL = "https://api.paperspace.io/deployments/updateDeployment/"
-    URL_V2 = "https://api.paperspace.io/deployments/v2/updateDeployment/"
+    URL = "https://api.paperspace.io/deployments/v2/updateDeployment/"
     COMMAND = ["deployments", "stop",
                "--id", "some_id"]
-    COMMAND_WITH_VPC_FLAG = ["deployments", "stop",
-                             "--id", "some_id",
-                             "--vpc"]
     REQUEST_JSON = {"isRunning": False, "id": u"some_id"}
     EXPECTED_STDOUT = "Deployment stopped\n"
 
@@ -442,22 +394,6 @@ class TestStopDeployment(object):
         result = runner.invoke(cli.cli, self.COMMAND)
 
         post_patched.assert_called_once_with(self.URL,
-                                             headers=EXPECTED_HEADERS,
-                                             json=self.REQUEST_JSON,
-                                             params=None,
-                                             files=None,
-                                             data=None)
-        assert result.output == self.EXPECTED_STDOUT
-        assert result.exit_code == 0
-
-    @mock.patch("gradient.cli.deployments.deployments_commands.http_client.requests.post")
-    def test_should_send_request_to_api_v2_when_vpc_flag_was_used(self, post_patched):
-        post_patched.return_value = MockResponse(status_code=204)
-
-        runner = CliRunner()
-        result = runner.invoke(cli.cli, self.COMMAND_WITH_VPC_FLAG)
-
-        post_patched.assert_called_once_with(self.URL_V2,
                                              headers=EXPECTED_HEADERS,
                                              json=self.REQUEST_JSON,
                                              params=None,
@@ -491,7 +427,7 @@ class TestStopDeployment(object):
         result = runner.invoke(cli.cli, command)
 
         assert result.output == self.EXPECTED_STDOUT, result.exc_info
-        post_patched.assert_called_once_with(self.URL_V2,
+        post_patched.assert_called_once_with(self.URL,
                                              headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
                                              json=self.REQUEST_JSON,
                                              params=None,
@@ -517,14 +453,10 @@ class TestStopDeployment(object):
 
 
 class TestDeleteDeployment(object):
-    URL = "https://api.paperspace.io/deployments/deleteDeployment"
-    URL_V2 = "https://api.paperspace.io/deployments/v2/deleteDeployment"
+    URL = "https://api.paperspace.io/deployments/v2/deleteDeployment"
 
     COMMAND = ["deployments", "delete",
                "--id", "some_id"]
-    COMMAND_WITH_VPC_FLAG = ["deployments", "delete",
-                             "--id", "some_id",
-                             "--vpc"]
     COMMAND_WITH_API_KEY = [
         "deployments", "delete",
         "--id", "some_id",
@@ -562,22 +494,6 @@ class TestDeleteDeployment(object):
         assert result.exit_code == 0
 
     @mock.patch("gradient.cli.deployments.deployments_commands.http_client.requests.post")
-    def test_should_send_request_to_api_v2_when_vpc_flag_was_used(self, post_patched):
-        post_patched.return_value = MockResponse(status_code=204)
-
-        runner = CliRunner()
-        result = runner.invoke(cli.cli, self.COMMAND_WITH_VPC_FLAG)
-
-        post_patched.assert_called_once_with(self.URL_V2,
-                                             headers=EXPECTED_HEADERS,
-                                             json=self.REQUEST_JSON,
-                                             params=None,
-                                             files=None,
-                                             data=None)
-        assert result.output == self.EXPECTED_STDOUT
-        assert result.exit_code == 0
-
-    @mock.patch("gradient.cli.deployments.deployments_commands.http_client.requests.post")
     def test_should_send_proper_data_with_custom_api_key_when_api_key_parameter_was_provided(self, post_patched):
         post_patched.return_value = MockResponse(status_code=204)
 
@@ -602,7 +518,7 @@ class TestDeleteDeployment(object):
         result = runner.invoke(cli.cli, command)
 
         assert result.output == self.EXPECTED_STDOUT, result.exc_info
-        post_patched.assert_called_once_with(self.URL_V2,
+        post_patched.assert_called_once_with(self.URL,
                                              headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
                                              json=self.REQUEST_JSON,
                                              params=None,
@@ -628,8 +544,7 @@ class TestDeleteDeployment(object):
 
 
 class TestDeploymentsUpdate(object):
-    URL = "https://api.paperspace.io/deployments/updateDeployment"
-    URL_V2 = "https://api.paperspace.io/deployments/v2/updateDeployment"
+    URL = "https://api.paperspace.io/deployments/v2/updateDeployment"
     EXPECTED_HEADERS_WITH_CHANGED_API_KEY = http_client.default_headers.copy()
     EXPECTED_HEADERS_WITH_CHANGED_API_KEY["X-API-Key"] = "some_key"
     BASIC_OPTIONS_COMMAND = [
@@ -637,11 +552,10 @@ class TestDeploymentsUpdate(object):
         "--id", "some_id",
         "--deploymentType", "tfserving",
     ]
-    BASIC_OPTIONS_COMMAND_WITH_USE_VPC_FLAG = [
+    BASIC_OPTIONS_UPDATE_COMMAND = [
         "deployments", "update",
         "--id", "some_id",
         "--deploymentType", "tfserving",
-        "--vpc",
     ]
     BASIC_OPTIONS_COMMAND_WITH_API_KEY = [
         "deployments", "update",
@@ -672,7 +586,6 @@ class TestDeploymentsUpdate(object):
         "--authUsername", "some_username",
         "--authPassword", "some_password",
         "--clusterId", "some_cluster_id",
-        "--vpc",
         "--apiKey", "some_key",
     ]
     COMMAND_WITH_OPTIONS_FILE = ["deployments", "update", "--optionsFile", ]  # path added in test
@@ -720,27 +633,10 @@ class TestDeploymentsUpdate(object):
         post_patched.return_value = MockResponse(self.RESPONSE_JSON_200)
 
         runner = CliRunner()
-        result = runner.invoke(cli.cli, self.BASIC_OPTIONS_COMMAND)
+        result = runner.invoke(cli.cli, self.BASIC_OPTIONS_UPDATE_COMMAND)
 
         assert result.output == self.EXPECTED_STDOUT, result.exc_info
         post_patched.assert_called_once_with(self.URL,
-                                             headers=EXPECTED_HEADERS,
-                                             json=self.BASIC_OPTIONS_REQUEST,
-                                             params=None,
-                                             files=None,
-                                             data=None)
-
-        assert result.exit_code == 0
-
-    @mock.patch("gradient.cli.deployments.deployments_commands.http_client.requests.post")
-    def test_should_send_request_to_api_v2_when_vpc_flag_was_used(self, post_patched):
-        post_patched.return_value = MockResponse(self.RESPONSE_JSON_200)
-
-        runner = CliRunner()
-        result = runner.invoke(cli.cli, self.BASIC_OPTIONS_COMMAND_WITH_USE_VPC_FLAG)
-
-        assert result.output == self.EXPECTED_STDOUT, result.exc_info
-        post_patched.assert_called_once_with(self.URL_V2,
                                              headers=EXPECTED_HEADERS,
                                              json=self.BASIC_OPTIONS_REQUEST,
                                              params=None,
@@ -757,7 +653,7 @@ class TestDeploymentsUpdate(object):
         result = runner.invoke(cli.cli, self.COMMAND_WITH_ALL_OPTIONS)
 
         assert result.output == self.EXPECTED_STDOUT, result.exc_info
-        post_patched.assert_called_once_with(self.URL_V2,
+        post_patched.assert_called_once_with(self.URL,
                                              headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
                                              json=self.ALL_OPTIONS_REQUEST,
                                              params=None,
@@ -790,7 +686,7 @@ class TestDeploymentsUpdate(object):
         result = runner.invoke(cli.cli, command)
 
         assert result.output == self.EXPECTED_STDOUT, result.exc_info
-        post_patched.assert_called_once_with(self.URL_V2,
+        post_patched.assert_called_once_with(self.URL,
                                              headers=self.EXPECTED_HEADERS_WITH_CHANGED_API_KEY,
                                              json=self.ALL_OPTIONS_REQUEST,
                                              params=None,